pyproject.toml
[project]
name = "limco"
version = "0.1.0"
description = "limco using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]>=0.148.0,<1.0.0",
    "python-dotenv>=1.0.0"
]

[project.scripts]
limco = "limco.main:run"
run_crew = "limco.main:run"
train = "limco.main:train"
replay = "limco.main:replay"
test = "limco.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
